+++ R/plot_type_ranef.R
@@ -1,6 +1,6 @@
 #' @importFrom lme4 ranef
 #' @importFrom tibble rownames_to_column
-#' @importFrom purrr map map_df
+#' @importFrom purrr map map_df map2
 #' @importFrom arm se.ranef
 #' @importFrom stats qnorm
 #' @importFrom forcats fct_reorder
@@ -13,6 +13,7 @@ plot_type_ranef <- function(model,
                             se,
                             tf,
                             sort.est,
+                            title,
                             axis.labels,
                             axis.lim,
                             grid.breaks,
@@ -172,7 +173,7 @@ plot_type_ranef <- function(model,
       tmp$facet <- grp.names[i]
       tmp$term <- factor(alabels)
       tmp$title <-
-        dplyr::if_else(facets, "", sprintf("Random effects of %s", grp.names[i]))
+        dplyr::if_else(facets, "Random effects", sprintf("Random effects of %s", grp.names[i]))
 
 
       # sort data frame, initial order
@@ -238,9 +239,10 @@ plot_type_ranef <- function(model,
       mydf <- list(mydf)
 
 
-    pl <- purrr::map(
+    pl <- purrr::map2(
       mydf,
-      function(x) {
+      1:length(mydf),
+      function(x, y) {
 
         # sort terms
         x$term <- forcats::fct_reorder(x$term, x$reihe)
@@ -253,13 +255,25 @@ plot_type_ranef <- function(model,
         names(labs) <- labs
 
 
+        # plot title
+
+        if (sjmisc::is_empty(title)) {
+          ptitle <- x[["title"]]
+        } else {
+          if (length(title) >= y)
+            ptitle <- title[y]
+          else
+            ptitle <- title
+        }
+
+
         # plot random effects
 
         plot_point_estimates(
           model = model,
           dat = x,
           tf = tf,
-          title = x[["title"]],
+          title = ptitle,
           axis.labels = labs,
           axis.title = NULL,
           axis.lim = axis.lim,